backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
pandas==2.2.1
pyparsing==3.1.1
pyjwt==2.7.0
scipy==1.13.0
scikit-learn==1.2.1
spotipy==2.23.0
numpy==1.26.4
flask-cors==4.0.0
python-dotenv==1.0.1